+++ core/urls.py
@@ -80,5 +80,10 @@ urlpatterns = [
     path('api/check/', skip_ga360(views_api.CheckView.as_view()), name='api-check'),
     path('api/data-service/comtrade/', skip_ga360(views_api.ComTradeDataView.as_view()), name='api-comtrade-data'),
     path('api/data-service/countrydata/', skip_ga360(views_api.CountryDataView.as_view()), name='api-country-data'),
+    path(
+        'api/data-service/tradebarrier/',
+        skip_ga360(views_api.TradeBarrierDataView.as_view()),
+        name='api-trade-barrier-data',
+    ),
 ]
 urlpatterns += redirects